Move Authorized and Registration modal to separate component
